pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "delta"
description="Delta Collectors"
authors = [{name = "Outside Open Developers", email = "developers@outsideopen.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
requires-python = ">=3.7"
dynamic = ["version"]

dependencies = [
    "camel-converter",
    "colorlog",
    "dbus-python",
    "persistQueue",
    "pyroute2",
    "python3-nmap",
    "requests",
]

[project.urls]
Source = "https://github.com/outsideopen/delta-collectors"

[project.optional-dependencies]
test = [
    "pytest >=7.2.1",
]

[project.scripts]
delta = "delta:main"

